refactor: consolidate model registry modules
