+++ tests/trainers/test_change.py
@@ -19,7 +19,7 @@ from torchgeo.datamodules import MisconfigurationException, OSCDDataModule
 from torchgeo.datasets import OSCD, RGBBandsMissingError
 from torchgeo.main import main
 from torchgeo.models import ResNet18_Weights
-from torchgeo.models._weights import WeightsEnum
+from torchgeo.models.registry import WeightsEnum
 from torchgeo.trainers import ChangeDetectionTask
