+++ tests/trainers/test_mae.py
@@ -15,7 +15,7 @@ from torch.nn import Module
 from torchgeo.datasets import SSL4EOS12
 from torchgeo.main import main
 from torchgeo.models import ViTSmall16_Weights
-from torchgeo.models._weights import WeightsEnum
+from torchgeo.models.registry import WeightsEnum
 from torchgeo.trainers import MAETask
