+++ tests/trainers/test_moco.py
@@ -14,7 +14,7 @@ from torch.nn import Module
 from torchgeo.datasets import SSL4EOS12, SeasonalContrastS2
 from torchgeo.main import main
 from torchgeo.models import ResNet18_Weights
-from torchgeo.models._weights import WeightsEnum
+from torchgeo.models.registry import WeightsEnum
 from torchgeo.trainers import MoCoTask
 
 from .test_classification import ClassificationTestModel
