+++ torchgeo/models/registry.py
@@ -1,13 +1,15 @@
 # Copyright (c) TorchGeo Contributors. All rights reserved.
 # Licensed under the MIT License.
 
-"""Register TorchGeo models and pretrained weights with timm."""
+"""TorchGeo model registry and pretrained weight definitions."""
 
+import enum
 from collections.abc import Callable
-from dataclasses import replace
+from dataclasses import dataclass, field, replace
 from typing import Any
 
 import timm
+import torch
 from timm.models import (
     PretrainedCfg,
     generate_default_cfgs,
@@ -16,7 +18,43 @@ from timm.models import (
 )
 from torch import nn
 
-from ._weights import WeightsEnum
+
+@dataclass
+class Weights(PretrainedCfg):
+    """Pretrained model configuration with TorchGeo metadata."""
+
+    transforms: nn.Module = field(default_factory=nn.Identity)
+    meta: dict[str, Any] = field(default_factory=dict)
+
+    def get_state_dict(
+        self, *args: Any, progress: bool = True, **kwargs: Any
+    ) -> dict[str, Any]:
+        """Download and return the pretrained state dict."""
+        if not isinstance(self.url, str):
+            raise ValueError('No URL is available for these weights')
+
+        state_dict: dict[str, Any] = torch.hub.load_state_dict_from_url(
+            self.url, *args, progress=progress, **kwargs
+        )
+        return state_dict
+
+
+class WeightsEnum(enum.Enum):
+    """Compatibility enum for TorchGeo pretrained configurations."""
+
+    def __getattr__(self, name: str) -> Any:
+        """Forward configuration attributes to the enum value."""
+        try:
+            value = object.__getattribute__(self, '_value_')
+        except AttributeError as ex:
+            raise AttributeError(name) from ex
+        return getattr(value, name)
+
+    def get_state_dict(self, *args: Any, **kwargs: Any) -> dict[str, Any]:
+        """Download and return the pretrained state dict."""
+        state_dict: dict[str, Any] = self.value.get_state_dict(*args, **kwargs)
+        return state_dict
+
 
 # timm looks for this mapping in the module that defines each entrypoint.
 default_cfgs: dict[str, PretrainedCfg] = {}
