+++ torchgeo/trainers/detection.py
@@ -23,7 +23,7 @@ from torchvision.ops import MultiScaleRoIAlign, feature_pyramid_network, misc
 from ..datamodules import BaseDataModule
 from ..datasets import RGBBandsMissingError, unbind_samples
 from ..datasets.utils import Sample
-from ..models._weights import WeightsEnum
+from ..models.registry import WeightsEnum
 from .base import BaseTask
 from .utils import GeneralizedRCNNTransformNoOp
