+++ torchgeo/trainers/simclr.py
@@ -23,7 +23,7 @@ import torchgeo.transforms as T
 
 from ..datasets.utils import Sample
 from ..models import get_weight
-from ..models._weights import WeightsEnum
+from ..models.registry import WeightsEnum
 from . import utils
 from .base import BaseTask
